requirements.txt
flask
google-cloud-secret-manager
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client